move read/write code to rw.go
